Update [11/11/2021]
- mod_seed.F90 : remove the maximum limit of trajectories per gridcell.

- mod_pos_tstep.F90 : include a missing reference to ijk in the calculation of the new position for the cases where there is no periodic zonal boundary.

- mod_vars.F90 and mod_init.F90 : include path to W file, if w is described explicitly.

- Update README.md

- Update runtest and the workflow to the new mod_pos_tstep.F90.

+++ Makefile
@@ -120,9 +120,9 @@ test :
 clean:
 
 	-rm -rf *.o *.mod *.out *.dSYM *.csv fort.* *.x *.in
-	-rm runtest.sh
 	-rm -rf _build
-	-rm $(RUNFILE)
+	test -s runtest.sh && rm runtest.sh || true
+	test -s $(RUNFILE) && rm $(RUNFILE) || true
 
 .PHONY : help
